Added history, changed prompt, added UI features
I enabled the chatbot to have access to the previous few messages sent. It should probably be 10 and not 5, I forgot that bot messages were saved too. I also changed the prompt to try and get better responses but needs tweaking. I also added in the UI features that accidentally didn't make this branch for quality of life.