Local Storage (#22)
* Created a chat-widget to be able to import into pages of the xl.

* Added chat-widget widget into shared to be used by pages of the xl.

* Adding chatbot button to all main pages

* Creating chatbot pop up feature

* Added some styling to match theme and added dummy message. Also added chat feature and way to distinguish user and bot messages

* Added fake demo chat

* Changed color one last time

* Okay last color tweak I promise

* Added auto scroll

* Added local storage

I stored the chat as a JSON file so that we can reuse it for the database more easily later. I also stored the current chat messages ID number to prevent two messages from being stored at the same ID upon reload.

* Added a justOpenedChat variable

Added variable to either auto scroll to the bottom if chat was just toggled or smooth scroll if user is currently messaging chatbot

* Updated chat send icon to material design

---------

Co-authored-by: Luke Allen <143120937+LukeAllen13@users.noreply.github.com>
Co-authored-by: Nicholas Cope <nicholascope2709@gmail.com>