Chatgpt integration (#24)
* Created a chat-widget to be able to import into pages of the xl.

* Added chat-widget widget into shared to be used by pages of the xl.

* Adding chatbot button to all main pages

* Creating chatbot pop up feature

* Added some styling to match theme and added dummy message. Also added chat feature and way to distinguish user and bot messages

* Added fake demo chat

* Changed color one last time

* Okay last color tweak I promise

* Added auto scroll

* Added local storage

I stored the chat as a JSON file so that we can reuse it for the database more easily later. I also stored the current chat messages ID number to prevent two messages from being stored at the same ID upon reload.

* Updated chat send icon to material design

* WIP implementation for chatbot API and backend OpenAI connection

* Almost working backend, fixed api issues

* Prettify dates returned by chatbot

* Added history, changed prompt, added UI features

I enabled the chatbot to have access to the previous few messages sent. It should probably be 10 and not 5, I forgot that bot messages were saved too. I also changed the prompt to try and get better responses but needs tweaking. I also added in the UI features that accidentally didn't make this branch for quality of life.

* Tweaked prompt, hard coded room time requests, added dummy rating

I tweaked the prompt a bunch and added a lot that I think can be hard coded from the reservation.py. Hard coded room time requests for better user experience and less feedback loop. I added a dummy rating system but I think it should be tweaked a litte.

* Chatbot responds with time ranges instead of start times, added close chat button

* Adjusted system prompt and message history limit

---------

Co-authored-by: Luke Allen <143120937+LukeAllen13@users.noreply.github.com>
Co-authored-by: ethanbonsall <ethanbonsall50@gmail.com>
Co-authored-by: Ethan Bonsall <123014150+ethanbonsall@users.noreply.github.com>